chatbot/backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Settings management
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# OpenAI
openai==1.58.1

# Google APIs
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.155.0

# Utilities
python-dotenv==1.0.1

